The internal handling of large integers (larger than 2^31-1 = 2147483647, which is the largest integer that can be represented in the signed 32 bit integers that PHP use) has been changed between PHP version 5.0.1 and 5.0.2.
The change manifests itself when the arguments to operators like >>
and % and functions like chr() is too large: previously this would go
unnoticed, but now the argument is clipped to 2^31-1.

For example, this means that 0xFFFFFFFF >> 24 gives 0xFF in version
5.0.1, but 0x7F in version 5.0.2 where it looks as if the left
argument is first clipped to 0x7FFFFFFF (the largest signed 32 bit
integer) and then shifted as expected.

Using base_convert() instead of the shift and modulo operations gives
correct results in both new and older versions of PHP 5.