Changed PelEntryTime to use a Julian Day Count internally instead of the more limited UNIX timestamp.  This enables use of the full range of Exif timestamps, all the way from 0000:01:01 to 9999:12:31.  The setValue() and getValue() methods still default to UNIX timestamps.
Actually using dates in year 0 is pushing things a bit too far.  This
is simply because there is no such thing as year 0: the first year was
year 1 and the year before that was year -1.  But apart from that all
dates from year 1 to year 9999 should work.

Also note, that by a lucky coincidence the illegal date '0000:00:00
00:00:00' is mapped to 0 which is mapped back to the string again.
Many cameras store that string when the date is unknown.

There might be some trouble with time zones... please let me know if
you spot any problems.