Fixing issue with copyright tag
Source: http://www.exiv2.org/tags.html

The copyright tag may contain ONE or TWO items. It is not guaranteed to be two items in the spec. When there is only one item, there is only a single NULL character, not a NULL separator with double-NULL termination. In this condition, $v[1] would not be set, and there for throw an error when trying to read this element within the array. 

Simple fix, check to see if the value is set, if not, set it to an empty string value.